pyproject.toml
[project]
name = "stargan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "black>=25.1.0",
    "datasets>=3.6.0",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "ipython>=9.3.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "onnxscript>=0.3.0",
    "pandas>=2.3.0",
    "pre-commit>=4.2.0",
    "torch>=2.7.1",
    "torchmetrics[image]>=1.7.3",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
    "wandb>=0.20.1",
]
